@@ -10,10 +10,11 @@
     "lib",
     "src",
     "bson.d.ts",
-    "etc/prepare.js"
+    "etc/prepare.js",
+    "vendor"
   ],
   "types": "bson.d.ts",
-  "version": "5.3.0",
+  "version": "5.4.0",
   "author": {
     "name": "The MongoDB NodeJS Team",
     "email": "dbx-node@mongodb.com"
@@ -26,7 +27,8 @@
   },
   "devDependencies": {
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
-    "@microsoft/api-extractor": "^7.34.7",
+    "@microsoft/api-extractor": "^7.35.1",
+    "@octokit/core": "^4.2.4",
     "@rollup/plugin-node-resolve": "^15.0.2",
     "@rollup/plugin-typescript": "^11.1.0",
     "@types/chai": "^4.3.5",
@@ -49,7 +51,6 @@
     "node-fetch": "^3.3.1",
     "nyc": "^15.1.0",
     "prettier": "^2.8.8",
-    "rimraf": "^5.0.0",
     "rollup": "^3.21.4",
     "sinon": "^15.0.4",
     "sinon-chai": "^3.7.0",
@@ -57,7 +58,7 @@
     "standard-version": "^9.5.0",
     "ts-node": "^10.9.1",
     "tsd": "^0.28.1",
-    "typescript": "^4.9.4",
+    "typescript": "^5.0.4",
     "typescript-cached-transpile": "0.0.6",
     "uuid": "^9.0.0",
     "v8-profiler-next": "^1.9.0"
@@ -85,7 +86,7 @@
       "types": "./bson.d.ts",
       "default": "./lib/bson.cjs"
     },
-    "react-native": "./lib/bson.cjs",
+    "react-native": "./lib/bson.rn.cjs",
     "browser": "./lib/bson.mjs"
   },
   "compass:exports": {
@@ -103,7 +104,7 @@
     "check:web": "WEB=true mocha test/node",
     "check:web-no-bigint": "WEB=true NO_BIGINT=true mocha test/node",
     "build:ts": "node ./node_modules/typescript/bin/tsc",
-    "build:dts": "npm run build:ts && api-extractor run --typescript-compiler-folder node_modules/typescript --local && rimraf 'lib/**/*.d.ts*' lib/parser lib/utils",
+    "build:dts": "npm run build:ts && api-extractor run --typescript-compiler-folder node_modules/typescript --local && node etc/clean_definition_files.cjs",
     "build:bundle": "rollup -c rollup.config.mjs",
     "build": "npm run build:dts && npm run build:bundle",
     "check:lint": "eslint -v && eslint --ext '.js,.ts' --max-warnings=0 src test && npm run build:dts && npm run check:tsd",
