@@ -1,6 +1,6 @@
 {
   "name": "mongodb",
-  "version": "5.5.0",
+  "version": "5.7.0",
   "description": "The official MongoDB driver for Node.js",
   "main": "lib/index.js",
   "files": [
@@ -25,7 +25,7 @@
     "email": "dbx-node@mongodb.com"
   },
   "dependencies": {
-    "bson": "^5.3.0",
+    "bson": "^5.4.0",
     "mongodb-connection-string-url": "^2.6.0",
     "socks": "^2.7.1"
   },
@@ -34,11 +34,19 @@
   },
   "peerDependencies": {
     "@aws-sdk/credential-providers": "^3.201.0",
+    "@mongodb-js/zstd": "^1.1.0",
+    "kerberos": "^2.0.1",
     "mongodb-client-encryption": ">=2.3.0 <3",
     "snappy": "^7.2.2"
   },
   "peerDependenciesMeta": {
     "@aws-sdk/credential-providers": {
+      "optional": true
+    },
+    "@mongodb-js/zstd": {
+      "optional": true
+    },
+    "kerberos": {
       "optional": true
     },
     "snappy": {
@@ -51,9 +59,10 @@
   "devDependencies": {
     "@iarna/toml": "^2.2.5",
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
-    "@microsoft/api-extractor": "^7.34.8",
+    "@microsoft/api-extractor": "^7.35.1",
     "@microsoft/tsdoc-config": "^0.16.2",
     "@mongodb-js/zstd": "^1.1.0",
+    "@octokit/core": "^4.2.1",
     "@types/chai": "^4.3.5",
     "@types/chai-subset": "^1.3.3",
     "@types/express": "^4.17.17",
@@ -88,10 +97,9 @@
     "sinon-chai": "^3.7.0",
     "snappy": "^7.2.2",
     "source-map-support": "^0.5.21",
-    "standard-version": "^9.5.0",
     "ts-node": "^10.9.1",
-    "tsd": "^0.27.0",
-    "typescript": "^4.9.5",
+    "tsd": "^0.28.1",
+    "typescript": "^5.0.4",
     "typescript-cached-transpile": "^0.0.6",
     "v8-heapsnapshot": "^1.2.0",
     "yargs": "^17.7.2"
@@ -110,6 +118,7 @@
     "build:dts": "npm run build:ts && api-extractor run && node etc/clean_definition_files.cjs",
     "build:docs": "./etc/docs/build.ts",
     "build:typedoc": "typedoc",
+    "build:nightly": "node ./.github/scripts/nightly.mjs",
     "check:bench": "node test/benchmarks/driverBench",
     "check:coverage": "nyc npm run test:all",
     "check:integration-coverage": "nyc npm run check:test",
@@ -120,6 +129,7 @@
     "check:tsd": "tsd --version && tsd",
     "check:dependencies": "mocha test/action/dependency.test.ts",
     "check:dts": "node ./node_modules/typescript/bin/tsc --noEmit mongodb.d.ts && tsd",
+    "check:search-indexes": "nyc mocha --config test/mocha_mongodb.json test/manual/search-index-management.spec.test.ts",
     "check:test": "mocha --config test/mocha_mongodb.json test/integration",
     "check:unit": "mocha test/unit",
     "check:ts": "node ./node_modules/typescript/bin/tsc -v && node ./node_modules/typescript/bin/tsc --noEmit",
@@ -127,6 +137,7 @@
     "check:adl": "mocha --config test/mocha_mongodb.json test/manual/atlas-data-lake-testing",
     "check:aws": "nyc mocha --config test/mocha_mongodb.json test/integration/auth/mongodb_aws.test.ts",
     "check:oidc": "mocha --config test/mocha_mongodb.json test/manual/mongodb_oidc.prose.test.ts",
+    "check:oidc-azure": "mocha --config test/mocha_mongodb.json test/integration/auth/mongodb_oidc_azure.prose.test.ts",
     "check:ocsp": "mocha --config test/manual/mocharc.json test/manual/ocsp_support.test.js",
     "check:kerberos": "nyc mocha --config test/manual/mocharc.json test/manual/kerberos.test.ts",
     "check:tls": "mocha --config test/manual/mocharc.json test/manual/tls_support.test.js",
@@ -137,7 +148,6 @@
     "fix:eslint": "npm run check:eslint -- --fix",
     "prepare": "node etc/prepare.js",
     "preview:docs": "ts-node etc/docs/preview.ts",
-    "release": "bash etc/check-remote.sh && standard-version -a -i HISTORY.md",
     "test": "npm run check:lint && npm run test:all",
     "test:all": "npm run check:unit && npm run check:test",
     "update:docs": "npm run build:docs -- --yes"
