@@ -1,7 +1,7 @@
 {
   "name": "mongoose",
   "description": "Mongoose MongoDB ODM",
-  "version": "7.2.2",
+  "version": "7.4.1",
   "author": "Guillermo Rauch <guillermo@learnboost.com>",
   "keywords": [
     "mongodb",
@@ -19,19 +19,19 @@
   ],
   "license": "MIT",
   "dependencies": {
-    "bson": "^5.3.0",
+    "bson": "^5.4.0",
     "kareem": "2.5.1",
-    "mongodb": "5.5.0",
+    "mongodb": "5.7.0",
     "mpath": "0.9.0",
     "mquery": "5.0.0",
     "ms": "2.1.3",
     "sift": "16.0.1"
   },
   "devDependencies": {
-    "@babel/core": "7.21.8",
-    "@babel/preset-env": "7.21.5",
-    "@typescript-eslint/eslint-plugin": "5.59.2",
-    "@typescript-eslint/parser": "5.59.2",
+    "@babel/core": "7.22.5",
+    "@babel/preset-env": "7.22.5",
+    "@typescript-eslint/eslint-plugin": "5.61.0",
+    "@typescript-eslint/parser": "5.61.0",
     "acquit": "1.3.0",
     "acquit-ignore": "0.2.1",
     "acquit-require": "0.1.1",
@@ -44,31 +44,32 @@
     "buffer": "^5.6.0",
     "cheerio": "1.0.0-rc.12",
     "crypto-browserify": "3.12.0",
-    "dotenv": "16.0.3",
+    "dotenv": "16.3.1",
     "dox": "1.0.0",
-    "eslint": "8.39.0",
+    "eslint": "8.44.0",
     "eslint-plugin-markdown": "^3.0.0",
     "eslint-plugin-mocha-no-only": "1.1.1",
     "express": "^4.18.1",
+    "fs-extra": "~11.1.1",
     "highlight.js": "11.8.0",
     "lodash.isequal": "4.5.0",
     "lodash.isequalwith": "4.4.0",
+    "markdownlint-cli2": "^0.8.1",
     "marked": "4.3.0",
-    "mkdirp": "^2.1.3",
+    "mkdirp": "^3.0.1",
     "mocha": "10.2.0",
     "moment": "2.x",
-    "mongodb-memory-server": "8.12.2",
+    "mongodb-memory-server": "8.13.0",
     "ncp": "^2.0.0",
     "nyc": "15.1.0",
     "pug": "3.0.2",
     "q": "1.5.1",
-    "sinon": "15.0.4",
+    "sinon": "15.2.0",
     "stream-browserify": "3.0.0",
     "tsd": "0.28.1",
-    "typescript": "5.0.4",
+    "typescript": "5.1.6",
     "uuid": "9.0.0",
-    "webpack": "5.81.0",
-    "fs-extra": "~11.1.1"
+    "webpack": "5.88.1"
   },
   "directories": {
     "lib": "./lib/mongoose"
@@ -98,9 +99,9 @@
     "docs:prepare:publish:6x": "npm run docs:checkout:6x && npm run docs:merge:6x && npm run docs:clean:stable && env DOCS_DEPLOY=true npm run docs:generate && npm run docs:move:6x:tmp && npm run docs:checkout:gh-pages && npm run docs:copy:tmp:6x",
     "docs:check-links": "blc http://127.0.0.1:8089 -ro",
     "lint": "eslint .",
-    "lint-js": "eslint . --ext .js",
+    "lint-js": "eslint . --ext .js --ext .cjs",
     "lint-ts": "eslint . --ext .ts",
-    "lint-md": "eslint . --ext .md",
+    "lint-md": "markdownlint-cli2 \"**/*.md\"",
     "build-browser": "(rm ./dist/* || true) && node ./scripts/build-browser.js",
     "prepublishOnly": "npm run build-browser",
     "release": "git pull && git push origin master --tags && npm publish",
